move test code analyzer and generator to a separate folder
