+++ src/main/kotlin/org/jetbrains/research/testspark/services/java/JavaTestCaseDisplayService.kt
@@ -37,8 +37,8 @@ import org.jetbrains.research.testspark.data.UIContext
 import org.jetbrains.research.testspark.display.TestCasePanelFactory
 import org.jetbrains.research.testspark.display.TopButtonsPanelFactory
 import org.jetbrains.research.testspark.display.utils.ReportUpdater
-import org.jetbrains.research.testspark.helpers.java.JavaTestAnalyzer
-import org.jetbrains.research.testspark.helpers.java.JavaTestGenerator
+import org.jetbrains.research.testspark.testmanager.java.JavaTestAnalyzer
+import org.jetbrains.research.testspark.testmanager.java.JavaTestGenerator
 import org.jetbrains.research.testspark.java.JavaPsiClassWrapper
 import org.jetbrains.research.testspark.langwrappers.PsiClassWrapper
 import org.jetbrains.research.testspark.services.CoverageVisualisationService
