+++ src/main/kotlin/org/jetbrains/research/testspark/testmanager/kotlin/KotlinTestGenerator.kt
@@ -1,6 +1,5 @@
-package org.jetbrains.research.testspark.helpers.java
+package org.jetbrains.research.testspark.testmanager.kotlin
 
-import com.intellij.lang.java.JavaLanguage
 import com.intellij.openapi.command.WriteCommandAction
 import com.intellij.openapi.diagnostic.Logger
 import com.intellij.openapi.project.Project
@@ -8,11 +7,12 @@ import com.intellij.psi.PsiDocumentManager
 import com.intellij.psi.PsiFile
 import com.intellij.psi.PsiFileFactory
 import com.intellij.psi.codeStyle.CodeStyleManager
+import org.jetbrains.kotlin.idea.KotlinLanguage
 import org.jetbrains.research.testspark.core.data.TestGenerationData
-import org.jetbrains.research.testspark.helpers.template.TestGenerator
+import org.jetbrains.research.testspark.testmanager.template.TestGenerator
 import java.io.File
 
-object JavaTestGenerator : TestGenerator {
+object KotlinTestGenerator : TestGenerator {
 
     private val log = Logger.getInstance(this::class.java)
 
@@ -26,7 +26,10 @@ object JavaTestGenerator : TestGenerator {
         otherInfo: String,
         testGenerationData: TestGenerationData,
     ): String {
-        var testFullText = printUpperPart(className, imports, packageString, runWith, otherInfo)
+        log.debug("[KotlinClassBuilderHelper] Generate code for $className")
+
+        var testFullText =
+            printUpperPart(className, imports, packageString, runWith, otherInfo)
 
         // Add each test (exclude expected exception)
         testFullText += body
@@ -36,24 +39,17 @@ object JavaTestGenerator : TestGenerator {
 
         testFullText.replace("\r\n", "\n")
 
-        /**
-         * for better readability and make the tests shorter, we reduce the number of line breaks:
-         *  when we have three or more sequential \n, reduce it to two.
-         */
+        // Reduce the number of line breaks for better readability
         return formatCode(project, Regex("\n\n\n(?:\n)*").replace(testFullText, "\n\n"), testGenerationData)
     }
 
     override fun formatCode(project: Project, code: String, generatedTestData: TestGenerationData): String {
         var result = ""
         WriteCommandAction.runWriteCommandAction(project) {
-            val fileName = generatedTestData.resultPath + File.separatorChar + "Formatted.java"
-            // create a temporary PsiFile
+            val fileName = generatedTestData.resultPath + File.separatorChar + "Formatted.kt"
+            // Create a temporary PsiFile
             val psiFile: PsiFile = PsiFileFactory.getInstance(project)
-                .createFileFromText(
-                    fileName,
-                    JavaLanguage.INSTANCE,
-                    code,
-                )
+                .createFileFromText(fileName, KotlinLanguage.INSTANCE, code)
 
             CodeStyleManager.getInstance(project).reformat(psiFile)
 
@@ -62,7 +58,7 @@ object JavaTestGenerator : TestGenerator {
 
             File(fileName).delete()
         }
-
+        log.info("Formatted result class: $result")
         return result
     }
 
@@ -77,22 +73,23 @@ object JavaTestGenerator : TestGenerator {
 
         // Add package
         if (packageString.isNotBlank()) {
-            testText += "package $packageString;\n"
+            testText += "package $packageString\n"
         }
 
-        // add imports
+        // Add imports
         imports.forEach { importedElement ->
             testText += "$importedElement\n"
         }
 
         testText += "\n"
 
-        // add runWith if exists
+        // Add runWith if exists
         if (runWith.isNotBlank()) {
-            testText += "@RunWith($runWith)\n"
+            testText += "@RunWith($runWith::class)\n"
         }
-        // open the test class
-        testText += "public class $className {\n\n"
+
+        // Open the test class
+        testText += "class $className {\n\n"
 
         // Add other presets (annotations, non-test functions)
         if (otherInfo.isNotBlank()) {
